requirements.txt
numpy
opencv-contrib-python
pathlib
click8
multiprocessing
Pillow
face-recognition
flask
